db populated

+++ Models/ContactInfo.cs
@@ -2,7 +2,7 @@ namespace CarServiceProject.Models
 {
     public class ContactInfo
     {
-        public int ContactInfoId { get; set; }
+        public long ContactInfoId { get; set; }
         public string Email { get; set; } = string.Empty;
         public string Phone { get; set; } = string.Empty;
     }
