+++ Models/CustomerFeedback.cs
@@ -2,7 +2,7 @@ using System.ComponentModel.DataAnnotations;
 
 public class CustomerFeedback
 {
-    public int CustomerFeedbackId { get; set; }
+    public long CustomerFeedbackId { get; set; }
     public string Name { get; set; } = string.Empty;
     public string? ImageUrl { get; set;}
     public string Comments { get; set; } = string.Empty;
