+++ Models/Employee.cs
@@ -2,11 +2,11 @@ namespace CarServiceProject.Models
 {
     public class Employee
     {
-        public int EmployeeId { get; set; }
+        public long EmployeeId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string? ImageUrl { get; set;}
         public string? Position { get; set; }
-        public int ContactInfoId { get; set; }
+        public long ContactInfoId { get; set; }
         public ContactInfo ContactInfo { get; set; } = default!;
     }
 }
